feat: move two card controllers to one file
