+++ src/controllers/card.js
@@ -1,6 +1,12 @@
 import httpClient from "../httpClient/httpClient";
 
-const postCard = async (card) => {
+export async function getCard() {
+  const randNum = Math.ceil(Math.random() * 100);
+  const { data: card } = await httpClient.get(`/posts/${randNum}`);
+  return card;
+}
+
+export async function postCard(card) {
   httpClient
     .post("/posts", {
       userId: 1,
@@ -12,6 +18,4 @@ const postCard = async (card) => {
     .catch((e) => {
       throw new Error(e.message);
     });
-};
-
-export default postCard;
+}
