+++ src/states/Dialog/Dialog.js
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from "react";
 import { v4 as uuidv4 } from "uuid";
 import { Dialog, DialogTitle, TextField, Button } from "@mui/material";
-import postCard from "../../controllers/postCard";
+import { postCard } from "../../controllers/card";
 import { Box } from "@mui/system";
 
 const CustomDialog = ({ isOpen, handleClose, cardList, setCardList }) => {
